feat(codes): generate export file asynchronously
Discounts can have hundreds of thousands of codes, so they need to be generated
out-of-band.

The "Download Codes" link in `/discounts` does the job of generating the export
files if it does not exist and downloading them if they do.

The GenerateCodesWorker does nothing more than delegating to CodeSet#generate.